feat: add tests functions, configurations, and fixes

+++ codebase/app/crud/default.py
@@ -12,13 +12,13 @@ class GenericCRUD(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
         self.model = model
 
     def read(self, db: Session, id: Any) -> Optional[ModelType] :
-        return db.query(self.model).get(id)
+        return db.get(self.model, id)
     
-    def list(self, db: Session) ->  Optional[List[ModelType]]:
-        return db.query(self.model).all()
+    def list(self, db: Session, skip: int = 0, limit: int = 100) ->  Optional[List[ModelType]]:
+        return db.query(self.model).offset(skip).limit(limit).all()
     
-    def list_by_foreign_key(self, db: Session, fk_name: str, fk_value: Any) -> Optional[List[ModelType]]:
-        return db.query(self.model).filter(getattr(self.model, fk_name) == fk_value).all()
+    def list_by_foreign_key(self, db: Session, fk_name: str, fk_value: Any, skip: int = 0, limit: int = 100) -> Optional[List[ModelType]]:
+        return db.query(self.model).filter(getattr(self.model, fk_name) == fk_value).offset(skip).limit(limit).all()
 
     def create(self, db: Session, instance: CreateSchemaType) -> ModelType:
         db_obj = self.model(**instance.model_dump())
@@ -33,7 +33,7 @@ class GenericCRUD(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
         if not db_obj:
             return None
 
-        for key, value in instance.dict(exclude_unset=True).items():
+        for key, value in instance.model_dump(exclude_unset=True).items():
             setattr(db_obj, key, value)
 
         db.commit()
@@ -42,10 +42,15 @@ class GenericCRUD(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
         return db_obj
 
     def delete(self, db: Session, id: Any) -> Any:
-        db_obj = db.query(self.model).get(id)
+        db_obj = db.get(self.model, id)
+        if not db_obj:
+            return None
         db.delete(db_obj)
         db.commit()
         return id
     
     def count(self, db: Session) -> int:
         return db.query(self.model).count()
+    
+    def count_by_foreign_key(self, db: Session, fk_name: str, fk_value: Any) -> int:
+        return db.query(self.model).filter(getattr(self.model, fk_name) == fk_value).count()
